Move the script harnesses out of the shared package root; drop dead Images/
test_env_base.py and test_actuators_base.py were hand-run smoke checks
living in the importable package root under test_ names. pytest collected
their functions as tests and errored on the unfillable env_class/cfg
arguments for any invocation that reached them — only the repository's
testpaths/norecursedirs settings were hiding it:

  $ pytest environments/shared/
  ERROR environments/shared/test_env_base.py::test_basic_functionality
  ...  fixture 'env_class' not found

Group them with view_model_base.py, which is the same kind of helper, into
environments/shared/harnesses/ under names pytest will not mistake for
tests:

  test_env_base.py       -> harnesses/env_smoke.py
  test_actuators_base.py -> harnesses/actuators.py
  view_model_base.py     -> harnesses/viewer.py

test_actuators() becomes run_actuator_test() and the internal test_*
checks become check_*; the per-species scripts are updated to match. The
coverage omit entry that existed to work around the old naming is
replaced by one covering the package.

Separately, remove the top-level Images/ directory: 25 MB whose three
GIFs are byte-identical to the copies under website/static/img/ that the
site and README actually reference. Nothing referred to Images/ except
the .dockerignore entry excluding it, also dropped.
