+++ .dockerignore
@@ -15,7 +15,6 @@ venv/
 node_modules/
 website/
 notebooks/
-Images/
 logs/
 models/
 *.zip
